requirements.txt
aiosqlite==0.20.0
cryptography==40.0.2
httpx==0.27.2
python-dotenv==1.1.0
Telethon==1.35.0
